+++ multimodal/src/autogluon/multimodal/__init__.py
@@ -10,3 +10,4 @@ from . import utils
 from . import optimization
 from . import predictor
 from .predictor import MultiModalPredictor, AutoMMPredictor
+from .utils import download
