+++ multimodal/src/autogluon/multimodal/constants.py
@@ -141,3 +141,6 @@ RAY_TUNE_CHECKPOINT = "ray_tune_checkpoint.ckpt"
 BEST_K_MODELS_FILE = "best_k_models.yaml"
 LAST_CHECKPOINT = "last.ckpt"
 MODEL_CHECKPOINT = "model.ckpt"
+
+# url
+S3_PREFIX = 's3://'
